feat(netutils)!: a conduit dial carries the fetch it guards
ConduitDial was a guard in shape only. Its one method handed back the
SocketAddr it wrapped, and zingo-price, a crate that has never heard of
the mixnet, turned that address into a socks5h:// proxy URL and built a
reqwest client around it. The type-level claim that a price fetch is
mixnet-only was discharged into a string one line into the body of the
function that made it.

The HTTP leg now lives below the seam. zingo-netutils gains a
socks5-fetch feature and a socks5_fetch module carrying
Socks5FetchError, the zingo-net-diag classification table, and
ConduitDial::fetch_text, which is the first operation a conduit performs
rather than describes. A caller reaches the wire by holding the guard,
and holding the guard is what keeps the conduit's in-flight count honest
for as long as the work runs.

zingo-price loses reqwest and rustls from its dependency set, along with
fetch_over, the proxy-URL formatting, the crypto-provider install, and
the classification table. What remains is the price data model and the
per-source parsers, and the crate names no network type at all.

The reqwest feature string moved across verbatim rather than being
re-derived. ADR 0021 pins verification to the compiled-in Mozilla bundle
on every platform, and the netutils standalone lockfile is where that
shim lives, so a native-roots reqwest there would reintroduce the
Android handshake failure recorded in zingolib#2531.

BREAKING: PriceError::RequestFailed carries a boxed Socks5FetchError
where it carried a reqwest::Error. The failure field is unchanged, and
the box keeps the enum the size it was.
