+++ zingo-netutils/Cargo.toml
@@ -38,6 +38,18 @@ nym-validator-client = { version = "1.21.3", optional = true, default-features =
     "http-client",
 ] }
 tokio-socks = { version = "0.5", optional = true }
+# The plain-HTTP leg the price fetch rides, dialed through the conduit's
+# SOCKS5 proxy. The feature string is copied verbatim from the price
+# crate it moved from: webpki roots rather than the platform verifier,
+# because ADR 0021 rules that verification is the compiled-in Mozilla
+# bundle on every platform, and no auto-selected provider, because
+# `crypto::ensure_default_crypto_provider` installs aws-lc-rs first.
+reqwest = { workspace = true, optional = true, default-features = false, features = [
+    "cookies",
+    "rustls-tls-webpki-roots-no-provider",
+    "json",
+    "socks",
+] }
 hyper-util = { version = "0.1.20", optional = true }
 tower = { version = "0.5", optional = true }
 tokio = { workspace = true, optional = true, features = [
@@ -72,6 +84,14 @@ socks5-transmit = [
     # chain walk rather than a private copy of it. Std-only and
     # zero-dependency, so it resolves in both this crate's own lockfile and
     # the parent workspace's without coupling either.
+    "dep:zingo-net-diag",
+]
+# The wallet-side SOCKS5-dialing HTTP fetch: one conduit-guarded request,
+# classified into a typed zingo-net-diag failure. Light in the same sense
+# as socks5-transmit — no nym-sdk — so it resolves in the parent
+# workspace's lockfile.
+socks5-fetch = [
+    "dep:reqwest",
     "dep:zingo-net-diag",
 ]
 nym = [
@@ -147,6 +167,7 @@ hyper-rustls = { version = "0.27", default-features = false, features = [
 hyper-util = "0.1.20"
 hyper = "1"
 rustls-pemfile = "2"
+reqwest = { version = "0.12.15", default-features = false }
 tokio = "1"
 http-body-util = "0.1.3"
 x509-parser = "0.18"
